add the ability to bench an existing compaction
This patch extends the compaction execute bench to allow executing
a compaction of sorted runs present in a db manifest. The runs are
specified in an env variable called COMPACTION as a a comma-seperated
list of SR IDs. If this environment variable is set, the tool will
compact those runs rather than the SSTs it generates. This is useful
to help debug/optimize slow compactions in a running db.